refactor: seperate history from queue
